% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_genes.R
\docType{data}
\name{case_genes}
\alias{case_genes}
\title{List of "case" mouse gene IDs}
\format{
A character list with 31 entries:
\describe{
    \item{geneID}{ensembl gene ID}
    }
}
\source{
Taliaferro Lab FMR1 localization
analysis. These genes depend on FMR1 for
efficient localization to neuronal
projections in CAD cells.
}
\usage{
case_genes
}
\description{
A list of ensembl mouse gene IDs to serve as
an example case gene set.
}
\keyword{datasets}
