% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CISBPRNA_hs_PWM.R
\docType{data}
\name{CISBPRNA_hs_PWM}
\alias{CISBPRNA_hs_PWM}
\title{CISBPRNA Homo sapiens PWM list}
\format{
A list of named matrices with 193 entries. The name is unique
  refering to both the RBP and its specific motif, the matrix columns
  represent positions, and the rows are the probability of occurance for each
  base "A", "C", "G", and "T". Each column sums to 1.
  \describe{
      \item{PWM}{probability matrix of a RBP motif}
}
}
\source{
\url{http://cisbp-rna.ccbr.utoronto.ca/bulk.php}
    Download Homo sapiens species archive
}
\usage{
CISBPRNA_hs_PWM
}
\description{
A list of position weight Matrices (PWM) of recognition motifs for RNA
binding proteins (RBPs) as documented by CISBPRNA for Homo sapiens.
}
\keyword{datasets}
