% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLengthorGC.R
\name{length_compare}
\alias{length_compare}
\alias{GC_compare}
\title{Compare the average Length or GC content of two DNAStringSet objects}
\usage{
length_compare(caseDNAStringSet, ctrlDNAStringSet)

GC_compare(caseDNAStringSet, ctrlDNAStringSet)
}
\arguments{
\item{caseDNAStringSet}{A DNAStringSet Object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}

\item{ctrlDNAStringSet}{A DNAStringSet Object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}
}
\value{
a summary table including the pvalue as calculated by
  \code{wilcox.test}, the mean values for each DNAStringSet, the log2 Fold
  change between those means, the effect size as calculated by
  \code{\link[effsize]{cliff.delta}}
}
\description{
\code{length_compare} and \code{GC_compare} compare the length and GC content
between case and control DNAStringSets respectively.
}
\section{Functions}{
\itemize{
\item \code{length_compare}: the average length of two different DNAStringSet objects
and returns a summary statistic table of the comparison.

\item \code{GC_compare}: the average GC content of two different DNAStringSet
objects and returns a summary statistic table of the comparison.
}}

\examples{
case_fasta <- Biostrings::readDNAStringSet("example_case.fa")
ctrl_fasta <- Biostrings::readDNAStringSet("example_ctrl.fa")
length_compare(case_fasta, ctrl_fasta)
GC_compare(case_fasta, ctrl_fasta)
}
\seealso{
\code{\link{getTxOut}}, \code{\link[Biostrings]{readDNAStringSet}},
  \code{\link[Biostrings]{letterFrequency}},
  \code{\link[effsize]{cliff.delta}}
}
