% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif_by_gene.R, R/Motif_compare.R
\name{cisBPRNA_by_gene}
\alias{cisBPRNA_by_gene}
\alias{RBNS_by_gene}
\alias{custom_PWM_by_gene}
\alias{cisBPRNA_compare}
\alias{RBNS_compare}
\alias{custom_PWM_compare}
\title{Count Motif occurances within each sequence of a DNAStringSet}
\usage{
cisBPRNA_by_gene(motif_path, RBPinfo, DNAStringset)

RBNS_by_gene(motif_path, DNAStringset)

custom_PWM_by_gene(PWM_list, DNAStringset)

cisBPRNA_compare(motif_path, RBPinfo, caseDNAStringset, ctrlDNAStringSet)

RBNS_compare(motif_path, caseDNAStringset, ctrlDNAStringSet)

custom_PWM_compare(PWM_list, caseDNAStringset, ctrlDNAStringSet)
}
\arguments{
\item{motif_path}{A path. To the directory containing Motif PWM txt files.}

\item{RBPinfo}{A path. To the RBPinfo table downloaded from cisbpRNA}

\item{PWM_list}{a named list of matrices. A list of named matrices where the
name is unique, and the matrix columns represent positions, and the rows
are the probability of each base "A", "C", "G", and "T". Each column should
sum to 1.}

\item{ctrlDNAStringSet}{A DNAStringSet Object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}

\item{DNAStringSet}{A DNAStringSet Object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}

\item{caseDNAStringSet}{A DNAStringSet Object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}
}
\value{
a dataframe containing motif counts across each sequences within the
  DNAStringSet

a summary table including the pvalue as calculated by
  \code{wilcox.test}, then corrected by \code{p.adjust} with the \code{method = "BH"}.
  The mean values for each DNAStringSet, and the log2 Fold
  change between those means
}
\description{
This function simply returns every motif counts for each gene. Begin with
folder containing PWM for motifs of intrest (downloaded from CisBPRNA called
"pwms_all_motifs") also need RBP info for the motifs (dowloaded from CisBPRNA
called "RBP_Information_all_motifs")
cisbpRNA:
\url{http://cisbp-rna.ccbr.utoronto.ca/bulk.php} ensure RBP info and PFMs are
selected for downloading an entire species archive

This function compares motif occurances between between case and control
DNAStringSets. Begin with a folder containing PWM for motifs of intrest
(downloaded from CisBPRNA called "pwms_all_motifs") also need RBP info for
the motifs (dowloaded from CisBPRNA called "RBP_Information_all_motifs")
cisbpRNA: \url{http://cisbp-rna.ccbr.utoronto.ca/bulk.php} ensure RBP info
and PFMs are selected for downloading an entire species archive
}
\section{Functions}{
\itemize{
\item \code{cisBPRNA_by_gene}: motif occurances for each sequence in a DNAStringSet.
cisbpRNA is an online library of RNA binding proteins and their motifs.
Motifs are derived from both direct and inferred binding.

\item \code{RBNS_by_gene}: motif occurances for each sequence in a DNAStringSet.
RNA bind and seq is an In vitro technique to define RBP binding sites.
These assays are performed wiht purified RBP and random synthetic RNAs.

\item \code{custom_PWM_by_gene}: custom PWM motif occurances for each sequence in a DNAStringSet.
This is useful for creating shorter PWM lists of special interest which will
count motifs much faster.

\item \code{cisBPRNA_compare}: motif occurances in case and ctrl DNAStringSets then
produces a summary table.
cisbpRNA is an online library of RNA binding proteins and their motifs.
Motifs are derived from both direct and inferred binding.

\item \code{RBNS_compare}: motif occurances in case and ctrl DNAStringSets then
produces a summary table.
RNA bind and seq is an In vitro technique to define RBP binding sites.
These assays are performed wiht purified RBP and random synthetic RNAs.

\item \code{custom_PWM_compare}: custom PWM motif occurances in case and ctrl DNAStringSets then
produces a summary table.
This is useful for creating shorter PWM lists of special interest which will
count motifs much faster.
}}

\examples{
case_fasta <- Biostrings::readDNAStringSet("example_case.fa")

cisBPRNA_by_gene("mydata/CisBPRNAdat/mm/pwms_all_motifs",
"mydata/CisBPRNAdat/mm/RBP_Information_all_motifs.txt", case_fasta)

RBNS_by_gene("mydata/RBNSdat/RBNS_PWMs", case_fasta)

custom_PWM_list <- readRDS("mydata/example_custom_PWM_list")
custom_PWM_by_gene(custom_PWM_list, case_fasta)

case_fasta <- Biostrings::readDNAStringSet("example_case.fa")
ctrl_fasta <- Biostrings::readDNAStringSet("example_ctrl.fa")

cisBPRNA_compare("mydata/CisBPRNAdat/mm/pwms_all_motifs",
"mydata/CisBPRNAdat/mm/RBP_Information_all_motifs.txt", case_fasta, ctrl_fasta)

RBNS_commpare("mydata/RBNSdat/RBNS_PWMs", case_fasta, ctrl_fasta)

custom_PWM_list <- readRDS("mydata/example_custom_PWM_list")
custom_PWM_compare(custom_PWM_list, case_fasta, ctrl_fasta)

}
\seealso{
\code{\link{getTxOut}}, \code{\link[Biostrings]{readDNAStringSet}},
  \code{\link[Biostrings]{countPWM}}, \code{\link{cisBPRNA_compare}},
   \code{\link{RBNS_compare}}, \code{\link{custom_PWM_compare}}

\code{\link{getTxOut}}, \code{\link[Biostrings]{readDNAStringSet}},
  \code{\link[Biostrings]{countPWM}}, \code{\link{cisBPRNA_by_gene}},
  \code{\link{RBNS_by_gene}}, \code{\link{custom_PWM_by_gene}}
}
