% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrl_genes.R
\docType{data}
\name{ctrl_genes}
\alias{ctrl_genes}
\title{List of "control" mouse gene IDs}
\format{
A character list with 7872 entries:
\describe{
    \item{geneID}{ensembl gene ID}
    }
}
\source{
Taliaferro Lab FMR1 localization
analysis. These genes are expessed in CAD
cells and do not depend on FMR1 for
efficient localization to neuronal
projections. This genes list excludes all
genes in case_genes.
}
\usage{
ctrl_genes
}
\description{
A list of ensembl mouse gene IDs to serve as
an example control gene set.
}
\keyword{datasets}
