% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_PWM_list.R
\docType{data}
\name{custom_PWM_list}
\alias{custom_PWM_list}
\title{custom PWM list examlpe}
\format{
A list of named matrices with 5 entries. The name is unique refering
  to the RBP, the matrix columns represent positions, and the rows are the
  probability of occurance for each base "A", "C", "G", and "T". Each column
  sums to 1.
   \describe{ \item{PWM}{probability matrix of a RBP motif} }
}
\source{
a subset of RBNS_PWM
}
\usage{
custom_PWM_list
}
\description{
A list of position weight Matrices (PWM) of recognition motifs for RNA
binding proteins (RBPs) to be used as an example
}
\keyword{datasets}
