% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_motif_from_kmer.R
\name{estimate_motif_from_kmer}
\alias{estimate_motif_from_kmer}
\title{Estimate Motif Occurance from Kmers}
\usage{
estimate_motif_from_kmer(
  kmer_list,
  motif_set,
  custom_motif_by_kmer_matrix = NULL
)
}
\arguments{
\item{kmer_list}{a character list of kmers}

\item{motif_set}{one of "CISBPRNA", "RBNS", or "custom"}

\item{custom_motif_by_kmer_matrix}{NULL unless motif_set = "custom". Use
output logical matrix from \code{relate_Kmer_PWM}.}
}
\value{
a summary table of motif matches in kmers including the pvalue as
  calculated by \code{phyper}, then corrected by \code{p.adjust} with the
  \code{method = "BH"}.
}
\description{
Because scanning for PWM matches across both the case and control
DNAStringSets can take quite a bit of time, this function estimates matches
from a kmer list. We recommend using significantly enriched or depleted kmers
as calculated by \code{kmer_compare}.
}
\examples{
enriched_sixmers <- c("AAGGAA", "ACACAC", "AGAAGG", "AGAGAG", "AGAGGG",
"AGGAAG", "AGGAGG", "AGGGAG", "CACACA", "GAAGGA", "GAGAAG", "GAGAGA",
"GAGGAG", "GAGGGA", "GAGGGG", "GGAAGG", "GGAGGA", "GGAGGG", "GGGAGG")
estimate_motif_from_kmer(enriched_sixmers, "RBNS")
estimate_motif_from_kmer(enriched_sixmers, "custom", )

}
