% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_Tx.R
\name{filter_Tx}
\alias{filter_Tx}
\title{Filter out undesired transcripts from gencode gff files}
\usage{
filter_Tx(gff, filter = TRUE, protein.coding = FALSE)
}
\arguments{
\item{gff}{A path. To the human or mouse gencode annotation gff file. The
most current annotations can be downloaded from:
\url{https://www.gencodegenes.org/human/} or
\url{https://www.gencodegenes.org/mouse/} There are two filter options, a
general filter and a protein coding filter.}

\item{filter}{A logical scalar. Should transcripts with undefined ends and
support levels higher than 2 (low confidence transcripts) be removed? The
default is to remove low confidence transcripts (\code{filter == TRUE})}

\item{protein.coding}{A logical scalar. Should all transcripts that lack a
protein coding tag be removed? The default is to keep all transcripts
regardless of protein coding status (\code{protein.coding == FALSE})}
}
\value{
The output from \code{filter_Tx()} is a TxDb object containing a gene
  model of the filtered transcripts This Txdb object is required for
  downstream functions. See
  \url{https://www.rdocumentation.org/packages/GenomicFeatures/versions/1.24.4/topics/TxDb-class}
  for more information
}
\description{
\code{filter_Tx} returns a filtered TxDb object
}
\examples{
# No filtering
hs_TxDb <- filter_Tx(system.file("extdata", "gencode.v33.annotation.gff3.gz", package = "FeatureReachR"), filter = FALSE, protein.coding = FALSE)
mm_TxDb <- filter_Tx(system.file("extdata", "gencode.vM20.annotation.gff3.gz", package = "FeatureReachR"), filter = FALSE, protein.coding = FALSE)

# Filter out low confidence transcripts
hs_filtered_TxDb <- filter_Tx(system.file("extdata", "gencode.v33.annotation.gff3.gz", package = "FeatureReachR"))
mm_filtered_TxDb <- filter_Tx(system.file("extdata", "gencode.vM20.annotation.gff3.gz", package = "FeatureReachR"))

# Filter out low confidence transcripts and non-coding transcripts
hs_coding_TxDb <- filter_Tx(system.file("extdata", "gencode.v33.annotation.gff3.gz", package = "FeatureReachR"), protein.coding = TRUE)
mm_coding_TxDb <- filter_Tx(system.file("extdata", "gencode.vM20.annotation.gff3.gz", package = "FeatureReachR"), protein.coding = TRUE)
}
