% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tx2gene.R
\name{gene2Tx}
\alias{gene2Tx}
\title{Get transcript list from gene list}
\usage{
gene2Tx(length_df, gene_list, seq_type)
}
\arguments{
\item{length_df}{A data frame. Created by \code{\link{make_longest_df}} or
\code{\link{make_median_df}}.}

\item{gene_list}{A character list. Must contain Ensembl gene IDs and be
compatible with the gff (the same species).}

\item{seq_type}{One of "whole", "CDS", "5pUTR", or "3pUTR".}
}
\value{
\code{gene2Tx} returns a character list of Ensembl transcript IDs
  associated with the length and sequence type of interest.
}
\description{
\code{gene2Tx} takes a gene list and returns a list of transcripts with the
longest or median length feature of interest.
}
\examples{
hs_filtered_TxDb <- filter_Tx("mydata/Gencodedat/gencode.v33.annotation.gff3.gz")
longest_hs <- make_longest_df(hs_filtered_TxDb)
hs_gene <- c("ENSG00000000419.12", "ENSG00000001167.14", "ENSG00000000938.13")
hs_tx <- gene2Tx(longest_hs, hs_gene, "UTR3")
}
\seealso{
\code{\link{filter_Tx}}, \code{\link{make_longest_df}},
  \code{\link{make_median_df}}
}
