% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctrl_tx.R
\name{get_ctrl_tx}
\alias{get_ctrl_tx}
\title{Get Control Transcript list from Case gene list}
\usage{
get_ctrl_tx(length_df, gene_list, seq_type)
}
\arguments{
\item{length_df}{A dataframe relating genes and transcritps by their length.
Should be the output from \code{make_longest} or \code{make_median}.}

\item{gene_list}{A character list of ensembl gene IDs.}

\item{seq_type}{One of "whole", "CDS", "5UTR" or "3UTR".}
}
\value{
a character list of ensembl  transcript IDs.
}
\description{
This will return transcripts representing all other genes not in
\code{gene_list} This requires creating a \code{length_df} see
\code{make_longest} or \code{make_median}. Generating a control transcript
list with this function isn't recommended as not all genes will be exprsesed
in a given system.It is best to use expression data from your system to
create a list of expressed control transcripts.
}
\examples{
mm_case_gene <- c("ENSMUSG00000097392", "ENSMUSG00000025607", "ENSMUSG00000030671", "ENSMUSG00000034764", "ENSMUSG00000116215", "ENSMUSG00000039556", "ENSMUSG00000066510", "ENSMUSG00000018160", "ENSMUSG00000114306", "ENSMUSG00000028277", "ENSMUSG00000037216", "ENSMUSG00000032299") #12genes
get_ctrl_tx(longest_mm, mm_case_gene, "UTR3")
}
