% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_by_gene.R
\name{kmer_by_gene}
\alias{kmer_by_gene}
\title{Count Kmers within each sequence of a DNAStringSet}
\usage{
kmer_by_gene(DNAStringSet, k)
}
\arguments{
\item{DNAStringSet}{A DNAStringSet Object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}

\item{k}{An Integer. The length of kmers to count. Values 3 < k < 8 are
recommended.}
}
\value{
a dataframe containing kmer counts across each sequences within the DNAStringSet
}
\description{
This function simply returns the counts of each kmer for each gene in a
DNAStringSet.
}
\examples{
case_fasta <- Biostrings::readDNAStringSet("example_case.fa")
case_fasta_kmer_counts <- kmer_by_gene(case_fasta, 6)
}
\seealso{
\code{\link{write_Sequence}}, \code{\link[Biostrings]{readDNAStringSet}},
  \code{\link[Biostrings]{oligonucleotideFrequency}}, \code{\link{kmer_compare}}
}
