% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_compare.R
\name{kmer_compare}
\alias{kmer_compare}
\title{Compare kmer content between two DNAStringSets}
\usage{
kmer_compare(caseDNAStringSet, ctrlDNAStringSet, k)
}
\arguments{
\item{caseDNAStringSet}{A DNAStringSet Object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}

\item{ctrlDNAStringSet}{A DNAStringSet Object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}
}
\value{
a summary table including the pvalue as calculated by
  \code{wilcox.test}, then corrected by \code{p.adjust} with the \code{method = "BH"}.
  The mean values for each DNAStringSet, and the log2 Fold
  change between those means
}
\description{
This function compares kmer content between between case and control
DNAStringSets
}
\examples{
case_fasta <- Biostrings::readDNAStringSet("example_case.fa")
ctrl_fasta <- Biostrings::readDNAStringSet("example_ctrl.fa")
kmer_compare(case_fasta, ctrl_fasta, 6)
}
\seealso{
\code{\link{write_Sequence}}, \code{\link[Biostrings]{readDNAStringSet}},
  \code{\link[Biostrings]{oligonucleotideFrequency}},\code{\link{kmer_by_gene}},
   \code{\link[effsize]{cliff.delta}}
}
