% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_length_df.R
\name{make_longest_df}
\alias{make_longest_df}
\alias{make_median_df}
\title{Create transcript length reference dataframe}
\usage{
make_longest_df(TxDb_gff)

make_median_df(TxDb_gff)
}
\arguments{
\item{TxDb_gff}{A TxDb object. Using \code{\link{filter_Tx}} to create this
TxDb object is recommended.}
}
\value{
Both \code{make_longest_df} and \code{make_median_df} create a
  dataframe relating Ensembl gene IDs and ensembl Transcript IDs. Because the
  longest CDS and longest 3'UTR  for a gene may not belong to the same
  transcript, there is a column for each sequence type.
}
\description{
Because creating these dataframes takes a minute or two and its likely a user
will be interested in multiple seq types (whole transcript, CDS, 5'UTR,
3'UTR) it is better to create and save this dataframe once.
}
\section{Functions}{
\itemize{
\item \code{make_longest_df}: creates a dataframe of the
transcripts with the longest features for each gene in a TxDb object

\item \code{make_median_df}: creates a dataframe of the
transcripts with the median length features for each gene in a TxDb object
}}

\examples{
hs_filtered_TxDb <- filter_Tx("mydata/Gencodedat/gencode.v33.annotation.gff3.gz")
longest_hs <- make_longest_df(hs_filtered_TxDb)
median_hs <- make_median_df(hs_filtered_TxDb)

}
\seealso{
\code{\link{filter_Tx}}, \code{\link{gene2tx}}
}
