% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif_by_gene.R
\name{motif_by_gene}
\alias{motif_by_gene}
\title{Count Motif occurances within each sequence of a DNAStringSet}
\usage{
motif_by_gene(PWM_list, DNAStringset)
}
\arguments{
\item{PWM_list}{a named list of position weight matrices. A list of named
matrices where the name is unique, and the matrix columns represent
positions, and the rows are the probability of each base "A", "C", "G", and
"T". Each column should sum to 1. FeatureReachR has three PWM_lists build in:
"CISBPRNA_mm_PWM", "CISBPRNA_hs_PWM", and "RBNS_PWM".}

\item{DNAStringSet}{A DNAStringSet Object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}
}
\value{
a dataframe containing motif counts across each sequences within the
  DNAStringSet
}
\description{
This function simply returns the number of times a PWM matches with each
sequence in a DNAStringSet object.
}
\examples{
case_fasta <- Biostrings::readDNAStringSet("example_case.fa")

motif_by_gene(CISBPRNA_mm_PWM, case_fasta)
motif_by_gene(RBNS_PWM, case_fasta)

}
\seealso{
\code{\link{write_Sequence}}, \code{\link[Biostrings]{readDNAStringSet}},
  \code{\link[Biostrings]{countPWM}}, \code{\link{motif_compare}}
}
