% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Length_GC_plot.R
\name{length_plot}
\alias{length_plot}
\alias{GC_plot}
\title{ggplot Wrappers for length/GC Violin Plots}
\usage{
length_plot(case, ctrl)

GC_plot(case, ctrl)
}
\arguments{
\item{case}{A DNAStringSet object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}

\item{ctrl}{A DNAStringSet object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}
}
\value{
a ggplot2 visual
}
\description{
This creates Violin plots comparing the mean length or GC of  case and
control groups. It displays sample size and the p value of a wilcox test.
}
\section{Functions}{
\itemize{
\item \code{length_plot}: length of all case and ctrl sequences as violin plots.

\item \code{GC_plot}: GC content of all case and ctrl sequences as violin plots.
}}

\examples{
case_fasta <- Biostrings::readDNAStringSet("example_case.fa")
ctrl_fasta <- Biostrings::readDNAStringSet("example_ctrl.fa")
length_plot(case_fasta, ctrl_fasta)
GC_plot(case_fasta, ctrl_fasta)
}
\seealso{
\code{get_length}, \code{get_GC}
}
