% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate_Kmer_PWM.R
\name{relate_kmer_PWM}
\alias{relate_kmer_PWM}
\title{Relate Kmers and PWM list}
\usage{
relate_kmer_PWM(k, PWM_list)
}
\arguments{
\item{k}{An integer. Recommended values of K are 3 < k < 8.}

\item{PWM_list}{a named list of position weight matrices. A list of named
matrices where the name is unique, and the matrix columns represent
positions, and the rows are the probability of each base "A", "C", "G", and
"T". Each column should sum to 1. RNAreachR has three PWM_lists build in:
"CISBPRNA_mm_PWM", "CISBPRNA_hs_PWM", and "RBNS_PWM".}
}
\value{
A logial matrix indicating if a motif can match to each kmer.
}
\description{
Create a logical matrix relating position weight matrices to kmers.
This creates the input for \code{estimate_motif_from_kmer}
These matrices take time to make. Logical matrices relating RBNS
and CISBPRNA motifs to 4-, 5-, 6- and 7-mers are included within
RNAreachR.
}
\examples{
#this table is already saved in RNAreachR
relate_Kmer_RBNS(4, RBNS_PWM)
relate_kmer_RBNS(6, custom_PWM_list)
}
\seealso{
\code{\link[Biostrings]{countPWM}}, \code{\link{motif_compare}}
}
