% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLengthorGC.R
\name{get_length}
\alias{get_length}
\alias{get_GC}
\title{Get length or GC of each sequence in a DNAStringSet object}
\usage{
get_length(DNAStringSet)

get_GC(DNAStringSet)
}
\arguments{
\item{DNAStringSet}{A DNAStringSet Object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}
}
\value{
A data frame with the length or GC content for each sequence in a
  DNAStringSet object
}
\description{
These functions return data frames containing the length or GC content for
each sequence in a DNAStringSet object.
}
\section{Functions}{
\itemize{
\item \code{get_length}: the length of each sequence in a
DNAStringSet object and returns a dataframe.

\item \code{get_GC}: both the length and GC content of each
sequence in a DNAStringSet object and returns a dataframe.
}}

\examples{
case_fasta <- Biostrings::readDNAStringSet("example_case.fa")
get_length(case_fasta)
get_GC(case_fasta)
}
\seealso{
\code{\link{getTxOut}}, \code{\link[Biostrings]{readDNAStringSet}},
  \code{\link[Biostrings]{letterFrequency}}
}
