% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_breakdown.R
\name{RNA_breakdown}
\alias{RNA_breakdown}
\title{Summarize RNA types in transcript list}
\usage{
RNA_breakdown(gff, tx_list)
}
\arguments{
\item{gff}{A path. To the human or mouse gencode annotation gff file. The
most current annotations can be downloaded from:
\url{https://www.gencodegenes.org/human/} or
\url{https://www.gencodegenes.org/mouse/} There are two filter options, a
general filter and a protein coding filter.}

\item{tx_list}{A character list. Must contain Ensembl transcript IDs and be
compatible with the gff (the same species)}
}
\value{
\code{RNA_breakdown} creates a summary table containing both the
number of transcripts categorized within an RNA type (count) and the
percent of the total set an RNA type represents (percent).
}
\description{
\code{RNA_breakdown} counts RNA types present in a transcript list and can be
used to better understand your case and control sets of transcripts. If you
instead have sets of genes, we recommend using expression data to find
expresed transcripts within your experiment. Alternatively, use
\code{\link{gene2tx()}} to create transcript sets from either
\code{\link{make_longest_df}} or \code{\link{make_median_df}}
}
\examples{
mm_tx <- c("ENSMUST00000159265.1", "ENSMUST00000027032.5", "ENSMUST00000130201.7", "ENSMUST00000157375.1")
RNA_breakdown("mydata/Gencodedat/gencode.vM20.annotation.gff3.gz", mm_tx)
}
\seealso{
\code{\link{gene2tx()}},
\code{\link{make_longest_df}}, \code{\link{make_median_df}}
}
