% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer2tree_logo_PWM.R
\name{kmer2logo}
\alias{kmer2logo}
\alias{kmer2tree}
\title{Group Kmers and Represent them as Motif Logos}
\usage{
kmer2logo(kmers, title = "")

kmer2tree(kmers, title = "")
}
\arguments{
\item{kmers}{character list of kmers of the length.}

\item{title}{string for plot titles defaults to no title.}
}
\value{
a tree, seqotif plots or a named list of PWMs
}
\description{
Kmers of interest can be collapsed or grouped by their levenshtein distance
(>2) to simplify enriched/depleted kmer outputs. These functions help visualize
how kmers are grouped as well as provide PWM which can be visualized using
\code{ggseqlogo}.
}
\section{Functions}{
\itemize{
\item \code{kmer2logo}: the tree and all seq logos together

\item \code{kmer2tree}: the tree of kmers with groups highlighted.
}}

\examples{
enriched_sixmers <- c("AAGGAA", "ACACAC", "AGAAGG", "AGAGAG", "AGAGGG",
"AGGAAG", "AGGAGG", "AGGGAG", "CACACA", "GAAGGA", "GAGAAG", "GAGAGA",
"GAGGAG", "GAGGGA", "GAGGGG", "GGAAGG", "GGAGGA", "GGAGGG", "GGGAGG")
kmer2logo(enriched_sixmers, "My Plot Title")
kmer2tree(enriched_sixmers, "My Tree Title")
kmer_PWMs <- kmer2PWM(enriched_sixmers)
kmer_PWMs

}
\seealso{
\code{kmer_compare}, \code{\link[ggseqlogo]{ggseqlogo}}, \code{\link[msa]{msa}}
}
