% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Length_GC_plot.R, R/kmer_motif_plot.R
\name{length_plot}
\alias{length_plot}
\alias{GC_plot}
\alias{kmer_plot}
\alias{motif_plot}
\title{ggplot Wrappers for length/GC Violin Plots}
\usage{
length_plot(case, ctrl)

GC_plot(case, ctrl)

kmer_plot(kmer_stats)

motif_plot(RBP_stats)
}
\arguments{
\item{case}{A DNAStringSet object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}

\item{ctrl}{A DNAStringSet object. Use
\code{\link[Biostrings]{readDNAStringSet}} on a fasta file to create.}
}
\value{
a ggplot2 visual

a ggplot2 visual
}
\description{
This creates Violin plots comparing the mean length or GC of  case and
control groups. It displays sample size and the p value of a wilcox test.

This creates Volcano plots comparing the enrichment of kmers or motifs in
case versus control groups. It plots the magnitude and significance of the
enrichment/depletion.
}
\section{Functions}{
\itemize{
\item \code{length_plot}: length of all case and ctrl sequences as violin plots.

\item \code{GC_plot}: GC content of all case and ctrl sequences as violin plots.

\item \code{kmer_plot}: all kmers with their magnitude (log2FoldChange) and
significance

\item \code{motif_plot}: all motifs with their magnitude (log2FoldChange) and
significance.
}}

\examples{
case_fasta <- Biostrings::readDNAStringSet("example_case.fa")
ctrl_fasta <- Biostrings::readDNAStringSet("example_ctrl.fa")
length_plot(case_fasta, ctrl_fasta)
GC_plot(case_fasta, ctrl_fasta)
case_fasta <- Biostrings::readDNAStringSet("example_case.fa")
ctrl_fasta <- Biostrings::readDNAStringSet("example_ctrl.fa")
kmer_stats <- kmer_compare(case_fasta, ctrl_fasta)
kmer_plot(kmer_stats)

RBNS_stats <- motif_compare(RBNS_PWM, case_fasta, ctrl_fasta)
motif_plot(RBNS_stats)
}
\seealso{
\code{get_length}, \code{get_GC}

\code{kmer_compare}, \code{motif_compare},
  \code{estimate_kmer_from_motif}
}
